requirements.txt
argparse
elasticsearch
openai
python-dotenv
langchain
langchain-community
langchain-openai
tiktoken
torch
git+https://github.com/huggingface/transformers.git
tqdm
unstructured
unstructured[all-docs]
GitPython
